tests/submission/Google/audio/loop/loop_property_exists.html

<!doctype html>
<html>
 <head>
  <title>audio.loop - existence test</title>
  <script src="../../../../resources/testharness.js"></script>
  <script src="../../../../resources/testharnessreport.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#dom-media-loop">spec reference</a></p>
  <audio id="audio">
  </audio>
  <div id="log"></div>
  <script>
test(function() {
 assert_true(
  "loop" in document.getElementById("audio"),
  "audioElement.loop property exists");
});
  </script>
 </body>
</html>
